Move StepUtilitiesConcern methods into IdvSession concern
Because two of the methods reference document_capture_session which is related to flow_session and may move into idv_session
And irs_reproofing? will probably be deleted soon
